IMPALA-2284: Disallow long (1<<30) strings in group_concat()
This is the first step to fix issues with large memory allocations. In
this patch, the built-in `group_concat` is no longer allowed to allocate
arbitraryly large strings and crash impala, but is limited to the upper
bound of possible allocations in Impala.

This patch does not perform any functional change, but rather avoids
unnecessary crashes. However, it changes the parameter type of
FindChunk() in MemPool to be a signed 64bit integer. This change allows
the mempool to allocate internally memory of more than one 1GB, but the
public interface of Allocate() is not changed, so the general limitation
remains. The reason for this change is as follows:

  1) In a UDF FunctionContext::Reallocate() would allocate slightly more
  than 512MB from the FreePool.
  2) The free pool tries to double this size to alloocate 1GB from the
  MemPool.
  3) The MemPool doubles the size again and overflows the signed 32bit
  integer in the FindChunk() method. This will then only allocate 1GB
  instead of the expected 2GB.

What happens is that one of the callers expected a larger allocation
than actually happened, which will in turn lead to memory corruption as
soon as the memory is accessed.

Reviewed-on: http://gerrit.cloudera.org:8080/858
Tested-by: Internal Jenkins
(cherry picked from commit 11e552426abc7d558f713502bab82b557c27b957)

+++ be/src/runtime/mem-pool.h
@@ -183,17 +183,17 @@ class MemPool {
   struct ChunkInfo {
     bool owns_data;  // true if we eventually need to dealloc data
     uint8_t* data;
-    int size;  // in bytes
+    int64_t size;  // in bytes
 
-    // number of bytes allocated via Allocate() up to but excluding this chunk;
-    // *not* valid for chunks > current_chunk_idx_ (because that would create too
-    // much maintenance work if we have trailing unoccupied chunks)
-    int cumulative_allocated_bytes;
+    /// number of bytes allocated via Allocate() up to but excluding this chunk;
+    /// *not* valid for chunks > current_chunk_idx_ (because that would create too
+    /// much maintenance work if we have trailing unoccupied chunks)
+    int64_t cumulative_allocated_bytes;
 
-    // bytes allocated via Allocate() in this chunk
-    int allocated_bytes;
+    /// bytes allocated via Allocate() in this chunk
+    int64_t allocated_bytes;
 
-    explicit ChunkInfo(int size);
+    explicit ChunkInfo(int64_t size);
 
     ChunkInfo()
       : owns_data(true),
@@ -231,12 +231,12 @@ class MemPool {
   // total allocated_bytes_ since it includes bytes in chunks that are not used.
   MemTracker* mem_tracker_;
 
-  // Find or allocated a chunk with at least min_size spare capacity and update
-  // current_chunk_idx_. Also updates chunks_, chunk_sizes_ and allocated_bytes_
-  // if a new chunk needs to be created.
-  // If check_limits is true, this call can fail (returns false) if adding a
-  // new chunk exceeds the mem limits.
-  bool FindChunk(int min_size, bool check_limits);
+  /// Find or allocated a chunk with at least min_size spare capacity and update
+  /// current_chunk_idx_. Also updates chunks_, chunk_sizes_ and allocated_bytes_
+  /// if a new chunk needs to be created.
+  /// If check_limits is true, this call can fail (returns false) if adding a
+  /// new chunk exceeds the mem limits.
+  bool FindChunk(int64_t min_size, bool check_limits);
 
   // Check integrity of the supporting data structures; always returns true but DCHECKs
   // all invariants.
